fixup! Add module l10n_ch_credit_control_dunning_fees_qr_report
