+++ l10n_ch_credit_control_dunning_fees_qr_report/models/__init__.py
@@ -1,3 +1,4 @@
 from . import credit_control_communication_report
 from . import account_move
+from . import ir_actions_report
 from . import res_partner_bank
